Move printer URL registration
Printer URL was previously sent with every request when using the
IppPrinter class. This commit moves that code to the constructor so that
a new printer instance is only created once using the URL passed to the
constructor.

A little refactoring was also done to allow the IPP printer to be
re-used for different application specific types of IPP printer (i.e:
label printer, manifest printer etc.)